+++ src/SimpleXLSX.php
@@ -136,16 +136,16 @@ class SimpleXLSX
     public $rowsExReader;
 
     /* @var SimpleXMLElement[] $sheets */
-    protected $sheets;
-    protected $sheetFiles = [];
-    protected $sheetMetaData = [];
+    public $sheets;
+    public $sheetFiles = [];
+    public $sheetMetaData = [];
+    public $sheetRels = [];
     // scheme
     public $styles;
-    protected $hyperlinks;
     /* @var array[] $package */
-    protected $package;
-    protected $sharedstrings;
-    protected $date1904 = 0;
+    public $package;
+    public $sharedstrings;
+    public $date1904 = 0;
 
 
     /*
@@ -186,8 +186,8 @@ class SimpleXLSX
             0x30 => '%1.0f');   //"##0.0E0";
         // }}}
     */
-    protected $errno = 0;
-    protected $error = false;
+    public $errno = 0;
+    public $error = false;
     /**
      * @var false|SimpleXMLElement
      */
@@ -206,18 +206,18 @@ class SimpleXLSX
             'comment' => '',
             'entries' => []
         ];
-        if ($filename && $this->_unzip($filename, $is_data)) {
-            $this->_parse();
+        if ($filename && $this->unzip($filename, $is_data)) {
+            $this->parseEntries();
         }
     }
 
-    protected function _unzip($filename, $is_data = false)
+    public function unzip($filename, $is_data = false)
     {
 
         if ($is_data) {
             $this->package['filename'] = 'default.xlsx';
             $this->package['mtime'] = time();
-            $this->package['size'] = $this->_strlen($filename);
+            $this->package['size'] = SimpleXLSX::strlen($filename);
 
             $vZ = $filename;
         } else {
@@ -287,23 +287,23 @@ class SimpleXLSX
 
             // Special case : value block after the compressed data
             if ($aP['GPF'] & 0x0008) {
-                $aP1 = unpack('V1CRC/V1CS/V1UCS', $this->_substr($vZ, -12));
+                $aP1 = unpack('V1CRC/V1CS/V1UCS', SimpleXLSX::substr($vZ, -12));
 
                 $aP['CRC'] = $aP1['CRC'];
                 $aP['CS'] = $aP1['CS'];
                 $aP['UCS'] = $aP1['UCS'];
                 // 2013-08-10
-                $vZ = $this->_substr($vZ, 0, -12);
-                if ($this->_substr($vZ, -4) === "\x50\x4b\x07\x08") {
-                    $vZ = $this->_substr($vZ, 0, -4);
+                $vZ = SimpleXLSX::substr($vZ, 0, -12);
+                if (SimpleXLSX::substr($vZ, -4) === "\x50\x4b\x07\x08") {
+                    $vZ = SimpleXLSX::substr($vZ, 0, -4);
                 }
             }
 
             // Getting stored filename
-            $aI['N'] = $this->_substr($vZ, 26, $nF);
+            $aI['N'] = SimpleXLSX::substr($vZ, 26, $nF);
             $aI['N'] = str_replace('\\', '/', $aI['N']);
 
-            if ($this->_substr($aI['N'], -1) === '/') {
+            if (SimpleXLSX::substr($aI['N'], -1) === '/') {
                 // is a directory entry - will be skipped
                 continue;
             }
@@ -313,9 +313,9 @@ class SimpleXLSX
             $aI['P'] = ($aI['P'] === '.') ? '' : $aI['P'];
             $aI['N'] = basename($aI['N']);
 
-            $vZ = $this->_substr($vZ, 26 + $nF + $mF);
+            $vZ = SimpleXLSX::substr($vZ, 26 + $nF + $mF);
 
-            if ($aP['CS'] > 0 && ($this->_strlen($vZ) !== (int)$aP['CS'])) { // check only if availabled
+            if ($aP['CS'] > 0 && (SimpleXLSX::strlen($vZ) !== (int)$aP['CS'])) { // check only if availabled
                 $aI['E'] = 1;
                 $aI['EM'] = 'Compressed size is not equal with the value in header information.';
             }
@@ -384,10 +384,6 @@ class SimpleXLSX
         return true;
     }
 
-    protected function _strlen($str)
-    {
-        return (ini_get('mbstring.func_overload') & 2) ? mb_strlen($str, '8bit') : strlen($str);
-    }
 
     public function error($num = null, $str = null)
     {
@@ -402,12 +398,7 @@ class SimpleXLSX
         return $this->error;
     }
 
-    protected function _substr($str, $start, $length = null)
-    {
-        return (ini_get('mbstring.func_overload') & 2) ? mb_substr($str, $start, ($length === null) ? mb_strlen($str, '8bit') : $length, '8bit') : substr($str, $start, ($length === null) ? strlen($str) : $length);
-    }
-
-    protected function _parse()
+    public function parseEntries()
     {
         // Document data holders
         $this->sharedstrings = [];
@@ -418,7 +409,7 @@ class SimpleXLSX
         if ($relations = $this->getEntryXML('_rels/.rels')) {
             foreach ($relations->Relationship as $rel) {
                 $rel_type = basename(trim((string)$rel['Type'])); // officeDocument
-                $rel_target = $this->_getTarget('', (string)$rel['Target']); // /xl/workbook.xml or xl/workbook.xml
+                $rel_target = SimpleXLSX::getTarget('', (string)$rel['Target']); // /xl/workbook.xml or xl/workbook.xml
 
                 if ($rel_type === 'officeDocument'
                     && $workbook = $this->getEntryXML($rel_target)
@@ -444,30 +435,35 @@ class SimpleXLSX
                         // Loop relations for workbook and extract sheets...
                         foreach ($workbookRelations->Relationship as $workbookRelation) {
                             $wrel_type = basename(trim((string)$workbookRelation['Type'])); // worksheet
-                            $wrel_path = $this->_getTarget(dirname($rel_target), (string)$workbookRelation['Target']);
-                            if (!$this->entryExists($wrel_path)) {
+                            $wrel_target = SimpleXLSX::getTarget(dirname($rel_target), (string)$workbookRelation['Target']);
+                            if (!$this->entryExists($wrel_target)) {
                                 continue;
                             }
 
-
                             if ($wrel_type === 'worksheet') { // Sheets
-                                if ($sheet = $this->getEntryXML($wrel_path)) {
+                                if ($sheet = $this->getEntryXML($wrel_target)) {
                                     $index = array_search((string)$workbookRelation['Id'], $index_rId, true);
                                     $this->sheets[$index] = $sheet;
-                                    $this->sheetFiles[$index] = $wrel_path;
+                                    $this->sheetFiles[$index] = $wrel_target;
+                                    $srel_d = dirname($wrel_target);
+                                    $srel_f = basename($wrel_target);
+                                    $srel_file = $srel_d . '/_rels/' . $srel_f  . '.rels';
+                                    if ($this->entryExists($srel_file)) {
+                                        $this->sheetRels[$index] = $this->getEntryXML($srel_file);
+                                    }
                                 }
                             } elseif ($wrel_type === 'sharedStrings') {
-                                if ($sharedStrings = $this->getEntryXML($wrel_path)) {
+                                if ($sharedStrings = $this->getEntryXML($wrel_target)) {
                                     foreach ($sharedStrings->si as $val) {
                                         if (isset($val->t)) {
                                             $this->sharedstrings[] = (string)$val->t;
                                         } elseif (isset($val->r)) {
-                                            $this->sharedstrings[] = $this->_parseRichText($val);
+                                            $this->sharedstrings[] = SimpleXLSX::parseRichText($val);
                                         }
                                     }
                                 }
                             } elseif ($wrel_type === 'styles') {
-                                $this->styles = $this->getEntryXML($wrel_path);
+                                $this->styles = $this->getEntryXML($wrel_target);
 
                                 // number formats
                                 $this->nf = [];
@@ -498,7 +494,7 @@ class SimpleXLSX
                                     }
                                 }
                             } elseif ($wrel_type === 'theme') {
-                                $this->theme = $this->getEntryXML($wrel_path);
+                                $this->theme = $this->getEntryXML($wrel_target);
                             }
                         }
 
@@ -602,10 +598,10 @@ class SimpleXLSX
     public function getEntryData($name)
     {
         $name = ltrim(str_replace('\\', '/', $name), '/');
-        $dir = $this->_strtoupper(dirname($name));
-        $name = $this->_strtoupper(basename($name));
+        $dir = SimpleXLSX::strtoupper(dirname($name));
+        $name = SimpleXLSX::strtoupper(basename($name));
         foreach ($this->package['entries'] as &$entry) {
-            if ($this->_strtoupper($entry['path']) === $dir && $this->_strtoupper($entry['name']) === $name) {
+            if (SimpleXLSX::strtoupper($entry['path']) === $dir && SimpleXLSX::strtoupper($entry['name']) === $name) {
                 if ($entry['error']) {
                     return false;
                 }
@@ -634,7 +630,7 @@ class SimpleXLSX
                     if ($entry['data'] === false) {
                         $entry['error'] = 2;
                         $entry['error_msg'] = 'Decompression of data failed.';
-                    } elseif ($entry['ucs'] > 0 && ($this->_strlen($entry['data']) !== (int)$entry['ucs'])) {
+                    } elseif ($entry['ucs'] > 0 && (SimpleXLSX::strlen($entry['data']) !== (int)$entry['ucs'])) {
                         $entry['error'] = 3;
                         $entry['error_msg'] = 'Uncompressed size is not equal with the value in header information.';
                     } elseif (crc32($entry['data']) !== $entry['crc']) {
@@ -654,10 +650,10 @@ class SimpleXLSX
     public function deleteEntry($name)
     {
         $name = ltrim(str_replace('\\', '/', $name), '/');
-        $dir = $this->_strtoupper(dirname($name));
-        $name = $this->_strtoupper(basename($name));
+        $dir = SimpleXLSX::strtoupper(dirname($name));
+        $name = SimpleXLSX::strtoupper(basename($name));
         foreach ($this->package['entries'] as $k => $entry) {
-            if ($this->_strtoupper($entry['path']) === $dir && $this->_strtoupper($entry['name']) === $name) {
+            if (SimpleXLSX::strtoupper($entry['path']) === $dir && SimpleXLSX::strtoupper($entry['name']) === $name) {
                 unset($this->package['entries'][$k]);
                 return true;
             }
@@ -665,32 +661,9 @@ class SimpleXLSX
         return false;
     }
 
-    protected function _strtoupper($str)
+    public static function strtoupper($str)
     {
         return (ini_get('mbstring.func_overload') & 2) ? mb_strtoupper($str, '8bit') : strtoupper($str);
-    }
-
-    protected function _getTarget($base, $target)
-    {
-        $target = trim($target);
-        if (strpos($target, '/') === 0) {
-            return $this->_substr($target, 1);
-        }
-        $target = ($base ? $base . '/' : '') . $target;
-        // a/b/../c -> a/c
-        $parts = explode('/', $target);
-        $abs = [];
-        foreach ($parts as $p) {
-            if ('.' === $p) {
-                continue;
-            }
-            if ('..' === $p) {
-                array_pop($abs);
-            } else {
-                $abs[] = $p;
-            }
-        }
-        return implode('/', $abs);
     }
 
     /*
@@ -701,30 +674,15 @@ class SimpleXLSX
     public function entryExists($name)
     {
  // 0.6.6
-        $dir = $this->_strtoupper(dirname($name));
-        $name = $this->_strtoupper(basename($name));
+        $dir = SimpleXLSX::strtoupper(dirname($name));
+        $name = SimpleXLSX::strtoupper(basename($name));
         foreach ($this->package['entries'] as $entry) {
-            if ($this->_strtoupper($entry['path']) === $dir && $this->_strtoupper($entry['name']) === $name) {
+            if (SimpleXLSX::strtoupper($entry['path']) === $dir && SimpleXLSX::strtoupper($entry['name']) === $name) {
                 return true;
             }
         }
 
         return false;
-    }
-
-    protected function _parseRichText($is = null)
-    {
-        $value = [];
-
-        if (isset($is->t)) {
-            $value[] = (string)$is->t;
-        } elseif (isset($is->r)) {
-            foreach ($is->r as $run) {
-                $value[] = (string)$run->t;
-            }
-        }
-
-        return implode('', $value);
     }
 
     public static function parseFile($filename, $debug = false)
@@ -736,8 +694,8 @@ class SimpleXLSX
     {
         $xlsx = new self();
         $xlsx->debug = $debug;
-        if ($xlsx->_unzip($filename, $is_data)) {
-            $xlsx->_parse();
+        if ($xlsx->unzip($filename, $is_data)) {
+            $xlsx->parseEntries();
         }
         if ($xlsx->success()) {
             return $xlsx;
@@ -783,56 +741,12 @@ class SimpleXLSX
 
     public function worksheet($worksheetIndex = 0)
     {
-
-
         if (isset($this->sheets[$worksheetIndex])) {
-            $ws = $this->sheets[$worksheetIndex];
-
-            if (!isset($this->hyperlinks[$worksheetIndex]) && isset($ws->hyperlinks)) {
-                $this->hyperlinks[$worksheetIndex] = [];
-                $sheet_rels = str_replace('worksheets', 'worksheets/_rels', $this->sheetFiles[$worksheetIndex]) . '.rels';
-                $link_ids = [];
-
-                if ($rels = $this->getEntryXML($sheet_rels)) {
-                    // hyperlink
-//                  $rel_base = dirname( $sheet_rels );
-                    foreach ($rels->Relationship as $rel) {
-                        $rel_type = basename(trim((string)$rel['Type']));
-                        if ($rel_type === 'hyperlink') {
-                            $rel_id = (string)$rel['Id'];
-                            $rel_target = (string)$rel['Target'];
-                            $link_ids[$rel_id] = $rel_target;
-                        }
-                    }
-                }
-                foreach ($ws->hyperlinks->hyperlink as $hyperlink) {
-                    $ref = (string)$hyperlink['ref'];
-                    if ($this->_strpos($ref, ':') > 0) { // A1:A8 -> A1
-                        $ref = explode(':', $ref);
-                        $ref = $ref[0];
-                    }
-//                      $this->hyperlinks[ $worksheetIndex ][ $ref ] = (string) $hyperlink['display'];
-                    $loc = (string)$hyperlink['location'];
-                    $id = (string)$hyperlink['id'];
-                    if ($id) {
-                        $href = $link_ids[$id] . ($loc ? '#' . $loc : '');
-                    } else {
-                        $href = $loc;
-                    }
-                    $this->hyperlinks[$worksheetIndex][$ref] = $href;
-                }
-            }
-
-            return $ws;
+            return $this->sheets[$worksheetIndex];
         }
         $this->error(6, 'Worksheet not found ' . $worksheetIndex);
 
         return false;
-    }
-
-    protected function _strpos($haystack, $needle, $offset = 0)
-    {
-        return (ini_get('mbstring.func_overload') & 2) ? mb_strpos($haystack, $needle, $offset, '8bit') : strpos($haystack, $needle, $offset);
     }
 
     /**
@@ -852,7 +766,7 @@ class SimpleXLSX
 
         $ref = (string)$ws->dimension['ref'];
 
-        if ($this->_strpos($ref, ':') !== false) {
+        if (SimpleXLSX::strpos($ref, ':') !== false) {
             $d = explode(':', $ref);
             $idx = $this->getIndex($d[1]);
 
@@ -905,7 +819,7 @@ class SimpleXLSX
             $col = $m[1];
             $row = $m[2];
 
-            $colLen = $this->_strlen($col);
+            $colLen = SimpleXLSX::strlen($col);
             $index = 0;
 
             for ($i = $colLen - 1; $i >= 0; $i--) {
@@ -970,7 +884,7 @@ class SimpleXLSX
 
             case 'inlineStr':
                 // Value is rich text inline
-                $value = $this->_parseRichText($cell->is);
+                $value = SimpleXLSX::parseRichText($cell->is);
 
                 break;
 
@@ -1026,12 +940,6 @@ class SimpleXLSX
         $t = (abs($d) > 0) ? ($d - 25569) * 86400 + round($t * 86400) : round($t * 86400);
 
         return (int)$t;
-    }
-
-    public function href($worksheetIndex, $cell)
-    {
-        $ref = (string)$cell['r'];
-        return isset($this->hyperlinks[$worksheetIndex][$ref]) ? $this->hyperlinks[$worksheetIndex][$ref] : '';
     }
 
     public function toHTML($worksheetIndex = 0)
@@ -1180,7 +1088,7 @@ class SimpleXLSX
             return false;
         }
         if (is_array($cell)) {
-            $cell = $this->_num2name($cell[0]) . $cell[1];// [3,21] -> D21
+            $cell = SimpleXLSX::num2name($cell[0]) . $cell[1];// [3,21] -> D21
         }
         if (is_string($cell)) {
             $result = $ws->sheetData->xpath("row/c[@r='" . $cell . "']");
@@ -1190,17 +1098,6 @@ class SimpleXLSX
         }
 
         return null;
-    }
-
-    protected function _num2name($num)
-    {
-        $numeric = ($num - 1) % 26;
-        $letter = chr(65 + $numeric);
-        $num2 = (int)(($num - 1) / 26);
-        if ($num2 > 0) {
-            return $this->_num2name($num2) . $letter;
-        }
-        return $letter;
     }
 
     public function getSheets()
@@ -1256,5 +1153,69 @@ class SimpleXLSX
     public function setDateTimeFormat($value)
     {
         $this->datetimeFormat = is_string($value) ? $value : false;
+    }
+
+    public static function getTarget($base, $target)
+    {
+        $target = trim($target);
+        if (strpos($target, '/') === 0) {
+            return SimpleXLSX::substr($target, 1);
+        }
+        $target = ($base ? $base . '/' : '') . $target;
+        // a/b/../c -> a/c
+        $parts = explode('/', $target);
+        $abs = [];
+        foreach ($parts as $p) {
+            if ('.' === $p) {
+                continue;
+            }
+            if ('..' === $p) {
+                array_pop($abs);
+            } else {
+                $abs[] = $p;
+            }
+        }
+        return implode('/', $abs);
+    }
+
+    public static function parseRichText($is = null)
+    {
+        $value = [];
+
+        if (isset($is->t)) {
+            $value[] = (string)$is->t;
+        } elseif (isset($is->r)) {
+            foreach ($is->r as $run) {
+                $value[] = (string)$run->t;
+            }
+        }
+
+        return implode('', $value);
+    }
+
+    public static function num2name($num)
+    {
+        $numeric = ($num - 1) % 26;
+        $letter = chr(65 + $numeric);
+        $num2 = (int)(($num - 1) / 26);
+        if ($num2 > 0) {
+            return SimpleXLSX::num2name($num2) . $letter;
+        }
+        return $letter;
+    }
+
+    public static function strlen($str)
+    {
+        return (ini_get('mbstring.func_overload') & 2) ? mb_strlen($str, '8bit') : strlen($str);
+    }
+
+    public static function substr($str, $start, $length = null)
+    {
+        return (ini_get('mbstring.func_overload') & 2) ? mb_substr($str, $start, ($length === null) ? mb_strlen($str, '8bit') : $length, '8bit') : substr($str, $start, ($length === null) ? strlen($str) : $length);
+    }
+
+    public static function strpos($haystack, $needle, $offset = 0)
+    {
+        return (ini_get('mbstring.func_overload') & 2) ? mb_strpos($haystack, $needle, $offset, '8bit') : strpos($haystack, $needle, $offset);
     }
 }