+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.10 (2023-04-27)
+* Cell comments returns in rowsEx
+* hyperlinks moved to rowsEx
+* all fields and methods public now
+* helper methods is static now
+
 ## 1.0.21 (2023-04-09)
 * fixed activeSheet(Tab)